Fix 'Cannot find path...' error on first run
Remove-Item shows an error if the path doesn't exist.

+++ z.psm1
@@ -504,7 +504,7 @@ function Save-CdCommandHistory($removeCurrentDirectory = $false) {
 function WriteHistoryToDisk() {
   $newList = GetAllHistoryAsText $global:history
   Out-File -InputObject $newList -FilePath "$cdHistory.tmp"
-  Remove-Item $cdHistory
+  if (Test-Path $cdHistory) { Remove-Item $cdHistory }
   Rename-Item -Path "$cdHistory.tmp" -NewName $cdHistory
 }
